+++ src/main/java/nextmainfocus/account/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package nextmainfocus.security;
+package nextmainfocus.account;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,7 +19,7 @@ import org.springframework.security.web.SecurityFilterChain;
 @EnableWebSecurity
 public class SecurityConfiguration {
 	@Autowired
-	SecurityService securityService;
+	AccountService securityService;
 
 	@Bean
 	public PasswordEncoder passwordEncoder() {
@@ -45,9 +45,9 @@ public class SecurityConfiguration {
 
 	@Bean
 	public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-		http.authorizeHttpRequests().requestMatchers("/healthcheck", "/api/v1/login/").permitAll().anyRequest()
+		http.authorizeHttpRequests().requestMatchers("/",
+				"/api/v1/login/").permitAll().anyRequest()
 				.authenticated().and().httpBasic();
-		http.csrf().disable();
 
 		return http.build();
 	}
